requirements.txt
python-fasthtml>=0.1.0
pytest>=7.0.0

